+++ main.cpp
@@ -2,8 +2,18 @@
 #include "interfaceadjuster.h"
 
 #include <QApplication>
+#include <QtGlobal>
+#include <QtDebug>
+#include <QTextStream>
+#include <QTextCodec>
+#include <QLocale>
+#include <QTime>
 
 Q_DECLARE_METATYPE(Screenshot)
+
+static QTextCodec *logCodec = NULL;
+static FILE *logStream = NULL;
+QString logFilePath = "debug.log";
 
 
 // TODO: design inconsitencies across platforms
@@ -11,8 +21,48 @@ Q_DECLARE_METATYPE(Screenshot)
 // TODO: multi-threading
 
 
+void customMessageOutput(QtMsgType type, const QMessageLogContext &context, const QString &msg)
+{
+    QByteArray localMsg = msg.toLocal8Bit();
+    QString fileName(context.file);
+    QByteArray file = logCodec->fromUnicode(fileName);
+    QTime time = QTime::currentTime();
+    QString formatedTime = time.toString("hh:mm:ss.zzz");
+    fprintf(logStream, "%s ", qPrintable(formatedTime));
+
+    switch (type) {
+    case QtDebugMsg:
+        fprintf(logStream, "Debug: %s (`%s:%u, %s`)\n", localMsg.constData(), file.constData(), context.line, context.function);
+        break;
+    case QtWarningMsg:
+        fprintf(logStream, "Warning: %s (`%s:%u, %s`)\n", localMsg.constData(), file.constData(), context.line, context.function);
+        break;
+    case QtInfoMsg:
+        fprintf(logStream, "Info: %s (`%s:%u, %s`)\n", localMsg.constData(), file.constData(), context.line, context.function);
+        break;
+    case QtCriticalMsg:
+        fprintf(logStream, "Critical: %s (`%s:%u, %s`)\n", localMsg.constData(), file.constData(), context.line, context.function);
+        break;
+    case QtFatalMsg:
+        fprintf(logStream, "Fatal: %s (`%s:%u, %s`)\n", localMsg.constData(), file.constData(), context.line, context.function);
+        abort();
+        break;
+    }
+    fflush(logStream);
+}
+
 int main(int argc, char *argv[])
 {
+    QByteArray envVar = qgetenv("QTDIR");       //  check if the app is ran via QT Creator
+
+    if (envVar.isEmpty())
+        logStream = _wfopen(logFilePath.toStdWString().c_str(), L"a");  // log to file, append mode
+    else
+        logStream = stderr;                     // print all errors to console
+
+    logCodec = QTextCodec::codecForName("UTF-8");
+    qInstallMessageHandler(customMessageOutput); // custom message handler for debugging
+
     QApplication a(argc, argv);
 
     QCoreApplication::setOrganizationName("Foyl");
@@ -81,9 +131,6 @@ int main(int argc, char *argv[])
     QObject::connect(&w, &MainWindow::clearCopyingStatusLabels,
                      &c, &Controller::clearCopyingStatusLabels);
 
-    QObject::connect(&w, &MainWindow::getScreenshotPathsPoolLength,
-                     &c, &Controller::returnScreenshotPathPoolLength);
-
     QObject::connect(&c, &Controller::sendProgressBarLength,
                      &w, &MainWindow::setProgressBarLength);
 
@@ -134,6 +181,9 @@ int main(int argc, char *argv[])
 
     QObject::connect(&w, &MainWindow::sendTreeWidgetPointer,
                      &c, &Controller::receiveTreeWidgetPointer);
+
+    QObject::connect(&c, &Controller::sendJpegQualityValue,
+                     &w, &MainWindow::setJpegQualityValue);
 
     w.bootStrap();
     c.bootStrap();